Refactor Vite/Node methods out of the general index.ts
