+++ packages/astro/src/core/errors/dev/index.ts
@@ -1,0 +1,2 @@
+export { collectErrorMetadata } from './utils.js';
+export { createCustomViteLogger, enhanceViteSSRError, getViteErrorPayload } from './vite.js';
